refactor: 목업데이터 위치 옮김
